refactor: move code watch code to repository
